Add ability to run manage from the serve command
This commit adds the ability to run zed manage in the same process as the
zed serve command. If the -manage flag is set with a duration greater
than 0 manage will run.

This commit also move the manage package to lake/api/manage package and
simplifies the logic for the monitor function.
