+++ cmd/zed/internal/lakemanage/scan.go
