+++ cmd/zed/manage/ztests/config.yaml
@@ -6,7 +6,7 @@ script: |
   zed create -q test3
   zed branch -use test2 -q live
   zed manage -config=inherit.yaml -log.path=inherit.log
-  zq -Z 'msg == "updating pool" | cut name, branch, interval | sort name' inherit.log > inherit.zson
+  zq -Z 'msg == "updating pool" | cut name, branch | sort name' inherit.log > inherit.zson
 
 inputs:
   - name: inherit.yaml
@@ -14,7 +14,6 @@ inputs:
       interval: 1s
       pools:
         - pool: test1
-          interval: 5s
         - pool: test2
           branch: "live"
 
@@ -23,18 +22,15 @@ outputs:
     data: | 
       {
           name: "test1",
-          branch: "main",
-          interval: 5
+          branch: "main"
       }
       {
           name: "test2",
-          branch: "live",
-          interval: 1
+          branch: "live"
       }
       {
           name: "test3",
-          branch: "main",
-          interval: 1
+          branch: "main"
       }
   - name: stderr
     data: ""
